短信签名模版接口返还参数更新

+++ volc-sdk-java/src/main/java/com/volcengine/model/beans/SmsTemplateInfo.java
@@ -26,7 +26,7 @@ public class SmsTemplateInfo {
     @JSONField(name = "Reason")
     String reason;
     @JSONField(name = "CreatedTime")
-    int createdTime;
+    long createdTime;
     @JSONField(name = "ShortUrlConfig")
     ShortUrlConfig shortUrlConfig;
 }
